nixos pam_env: use conffile instead of envfile
Change the internals of how the PAM module "pam_env" is configured.
So far pam_env read simple KEY=VALUE pairs from a file specified with the
option "envfile". Now, pam_env is switched to use the option "conffile" which
reads a file that allows more complex configurations. For example the file
read by "conffile" can contain special variables that are expanded by pam_env.

The effect of the switch to using "conffile" should be purely internal. On the
one hand the pam_env option "readenv=0" is used to prevent pam_env from loading
the default environment file at "/etc/environment" that would pam_env would
fall back to now. On the other hand the mechanism of creating the KEY=VALUE
pairs for the "envfile" option is adapted to fit the new format of the
"conffile" option. So there should only be a visible effect if a VALUE contained
any special variables like "@{HOME}" or "@{SHELL}" that are expanded now by
pam_env.